requirements.txt
flask
python-dotenv
openai
tiktoken
markdown
gunicorn
faiss-cpu
numpy
sentence-transformers
torch

